requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.1
python-dotenv==1.0.1
requests==2.31.0
tenacity==8.2.3
urllib3==2.2.1
